backend/During/requirements.txt
torch
torchvision
torchaudio
transformers
sentence-transformers

Pillow
pandas
requests
tqdm
pymongo
beautifulsoup4
openpyxl
tabulate 
fastapi
dotenv
python-multipart
uvicorn



scikit-learn

pyjwt